Obsolete files moved and hook created.
