+++ src/components/AddCocktail/AddCocktail.tsx
@@ -1,8 +1,9 @@
-import { useMutation, gql, useQuery } from '@apollo/client';
+import { useMutation, gql } from '@apollo/client';
 import { Button, Form, Input, message, Select } from 'antd';
 import 'antd/dist/antd.css';
 import { AddForm, AddSpace } from './AddCocktailStyle';
 import { CocktailDetails } from '../Type/CocktailDetailsType';
+import { useGetCocktails } from '../hooks/useGetCocktails';
 
 const { Option } = Select;
 
@@ -18,23 +19,6 @@ const NEW_COCKTAIL = gql`
     }
   }
 `;
-
-const GET_COCKTAILS = gql`
-    query getFavorites {
-        cocktails {
-            id
-            name
-            howTo
-            ingredients
-            image
-            favorite
-        }
-    }
-`;
-
-type CocktailDetailsArray = {
-  cocktails: CocktailDetails[]
-};
 
 type AddCocktailFormValue = {
   name: string,
@@ -54,8 +38,7 @@ type AddCocktailInput = {
 
 export const AddCocktail = () => {
 
-  const { refetch } = useQuery<CocktailDetailsArray | null>(GET_COCKTAILS);
-
+  const { refetch } = useGetCocktails();
 
   const [form] = Form.useForm();
   const [newCocktail] = useMutation<{ newCocktail: CocktailDetails }, { input: AddCocktailInput }>(NEW_COCKTAIL);
@@ -68,7 +51,7 @@ export const AddCocktail = () => {
         image: values.image,
         favorite: JSON.parse(values.favorite)
       }}});
-      await refetch(GET_COCKTAILS);
+      await refetch();
       form.resetFields();
       message.success(`${values.name} has been added.`);
       console.log('Success:', values);
@@ -87,7 +70,6 @@ export const AddCocktail = () => {
     form={form}
     labelCol={{ span: 6 }}
     wrapperCol={{ span: 16 }}
-    // initialValues={{ name: "", howTo: "", ingredients: "", image: ""}}
     onFinish={(values) => onFinish(values as AddCocktailFormValue)}
     onFinishFailed={onFinishFailed}
     autoComplete="off"
